improved script, add remove archive prompt

+++ luxtronik_firmware.sh
@@ -15,13 +15,25 @@ if [[ "$filename" == *"wpreg."* ]]; then
 	curl -s "https://www.heatpump24.com/software/fetchSoftware.php?softwareID=$softwareID" --output "$filename"
 	echo "Downloaded as $filename"
 	echo "Decompress downloaded firmware"
-	python extract_V1_firmware.py $filename
-	rm $filename
+	./extract_V1_firmware.py $filename
+	read -p "Do you want to delete the archive file? [yN] " yn
+	if [[ "$yn" == "y" ]]; then
+		rm $filename
+	fi
 	exit
 else
 	curl -s "https://www.heatpump24.com/software/fetchSoftware.php?softwareID=$softwareID" --output "$filename.gz"
 	echo "Downloaded as $filename.gz"
 fi
+
+# Get the filenames before extracting
+files=($(gzip -qdkc $filename.gz | tar tf -))
+
+gz=${files[0]}
+md5=${files[1]}
+
+echo $gz
+echo $md5
 
 echo "Decompress downloaded firmware"
 gzip -qdkc "$filename.gz" | tar xf -
@@ -29,7 +41,7 @@ gzip -qdkc "$filename.gz" | tar xf -
 rm "$filename.gz"
 
 echo "Verify Checksum"
-find . -name 'home.*.md5' -print0 | xargs -0 md5sum -c
+md5sum -c $md5
 
 if [ $? -ne 0 ]; then
 	echo "Error, checksum does not match!"
@@ -37,17 +49,20 @@ if [ $? -ne 0 ]; then
 fi
 
 # Delete md5 sum file, not needed anymore
-rm home.*.md5
+rm $md5
 
 # Rename tar file to make tar happy
-filename=$(find . -name 'home.wp2reg*')
-mv $filename "$filename.tar"
+tar="$gz.tar"
+mv $gz "$tar"
 
 # Make a new directory for untared firmware
 mkdir firmware >/dev/null 2>&1
 
 echo "Extract firmware into ./firmware"
-tar xf "$filename.tar" -C firmware
+tar xf $tar -C firmware
 
 # Delete tar ball
-rm "$filename.tar"
+read -p "Do you want to delete the archive file? [yN] " yn
+if [[ "$yn" == "y" ]]; then
+	rm $tar
+fi
